requirements.txt
# Core Backend
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
python-dotenv==1.0.0

# Database
sqlite3  # Built-in

# Resume Parsing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# AI/ML - Semantic Embeddings
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3

# NLP Processing
nltk==3.8.1
spacy==3.7.2

# Email
smtplib  # Built-in
email  # Built-in

# Async Processing
celery==5.3.4
redis==5.0.1

# Utilities
python-dateutil==2.8.2
jinja2==3.1.2

